requirements.txt
streamlit
pandas
groq
tabulate
openpyxl
python-docx